Moved AssignedDownload to a separated file
