+++ src/Stratis.Bitcoin/BlockPulling2/BlockPuller.cs
@@ -646,7 +646,6 @@ namespace Stratis.Bitcoin.BlockPulling2
                             JobId = downloadJob.Id,
                             AssignedTime = this.dateTimeProvider.GetUtcNow(),
                             BlockHeight = header.Height,
-                            BlockHash = hashToAssign
                         });
 
                         this.logger.LogTrace("Block '{0}' was assigned to peer ID {1}.", hashToAssign, peerId);
@@ -978,33 +977,6 @@ namespace Stratis.Bitcoin.BlockPulling2
 
             /// <summary>Hashes of blocks that are to be downloaded.</summary>
             public List<uint256> Hashes;
-        }
-
-        /// <summary>Represents a single download assignment to a peer.</summary>
-        private struct AssignedDownload
-        {
-            /// <summary>Unique identifier of a job to which this assignment belongs.</summary>
-            public int JobId;
-
-            /// <summary>Id of a peer that was assigned to deliver a block.</summary>
-            public int PeerId;
-            
-            /// <summary>Time when download was assigned to a peer.</summary>
-            public DateTime AssignedTime;
-
-            /// <summary>Height of a block associated with this assignment.</summary>
-            public int BlockHeight;
-            
-            /// <summary>Hash or the requested block.</summary>
-            public uint256 BlockHash;
-
-            public LinkedListNode<AssignedDownload> LinkedListNode;
-
-            /// <inheritdoc />
-            public override string ToString()
-            {
-                return string.Format("{0}:{1},{2}:{3},{4}:{5}", nameof(this.JobId), this.JobId, nameof(this.PeerId), this.PeerId, nameof(this.BlockHeight), this.BlockHeight);
-            }
         }
     }
 }
